% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\docType{data}
\name{Annotation}
\alias{Annotation}
\title{Annotation data}
\format{A 36226 x 12 data frame .}
\usage{
data(Annotation)
}
\description{
Agilent array mouse annotation data with probe and gene information.
}
\details{
#' @docType data
}
\examples{
data(Annotation)
head(Annotation)
}
\keyword{datasets}
