% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData.R
\docType{data}
\name{pData}
\alias{pData}
\title{pData}
\format{
A 91 x 9 data frame .
}
\usage{
data(pData)
}
\description{
Internal SAGA sample array with information on array/file name, vector and some other useful stuff like color for plots etc.
}
\examples{
data(pData)
head(pData)
}
\keyword{datasets}
