% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_batch.R
\name{saga_batch}
\alias{saga_batch}
\title{Batch correction of joint SAGA data sets.}
\usage{
saga_batch(
  matrix.SAGA.qn,
  matrix.test.qn,
  rawdata,
  pData.joint,
  plotnumber = 2
)
}
\arguments{
\item{matrix.SAGA.qn}{quantile normalized saga matrix}

\item{matrix.test.qn}{quantile normalized test matrix}

\item{rawdata}{SAGA raw data}

\item{pData.joint}{matrix with combined SAGA data and user samples.}

\item{plotnumber}{default=2 for PCA plot with samples in black; alternative: 1 for t-SNE plot}
}
\value{
\code{eset.batch} batch corrected data set.
}
\description{
\code{saga_batch} removes batch effects from the joint data set. For each sample, the user must specify a batch number in the
"SAGA_USER_Samples.txt" file. If no batch numbers are available - you'll still have to provide a number for the samples. You can use
any integer number beginning with 1. The batch correction of the data set relies on the ComBat algorithm/package.
}
