% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gentargets.R
\name{saga_gentargets}
\alias{saga_gentargets}
\title{Create SAGA samples target file.}
\usage{
saga_gentargets(smplpath)
}
\arguments{
\item{smplpath}{path to the saga data folder with the user samples.}
}
\value{
\code{targets} default target file for user samples.
}
\description{
\code{saga_gentargets} will automatically create a "SampleInformation.txt" file (SIF) if none is provided. If you chose to use this function, a folder
with sample *.txt files is needed. Each textfile must contain the full Agilent microarray information.
}
\details{
\cr PLEASE NOTE: When the targets file is generated it cannot know about your sample batches. Batch defaults to No. for each sample is 11. You
can use any other number to indicate your batches BUT you'll have to change the numbers manually (i.e, in Excel or in a text editor).
}
