% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gsea.R
\name{saga_gsea}
\alias{saga_gsea}
\title{Geneset Enrichment Analysis (GSEA) of SAGA data.}
\usage{
saga_gsea(smplpath, saveResults = 0)
}
\arguments{
\item{smplpath}{path to the saga data folder with the user samples.}

\item{saveResults}{can be 1 for yes and 0 for no.}
}
\value{
\code{result} GSEA result - will also be saved into the sample folder.
}
\description{
\code{saga_gsea} performs batch-wise GSEA for SAGA samples.
}
