% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_import.R
\name{saga_import}
\alias{saga_import}
\title{Import user data for SAGA model building}
\usage{
saga_import(smplpath, showjoint = 0)
}
\arguments{
\item{smplpath}{path to the saga data folder with the user samples.}

\item{showjoint}{can be 1 or 0. In case of 1 it will show a boxplot of the unnormalised data (model + user data). Default is 0.}
}
\value{
\code{matrix.joint} n-by-m matrix of joint data (SAGA data + user samples).

\code{pData.joint} joint target matrix of both SAGA data and user samples.

\code{pData.user} target matrix of user samples.

\code{filelist} list of user array names.

\code{pData} phenotype target matrix of SAGA data

\code{SAGA_Data} SAGA core data; needs to loaded from sagadata
}
\description{
The \code{saga_import} function loads SAGA data from textfiles in a user specified path and prepares the necessary objects for later
classification of unknown arrays.
}
\details{
In addition to the sample files, a user-defined targets file must be provided. It contains the necessary sample names, batch
information etc. You can use the \code{saga_targets} function to create a targets file for you. However, in any case, the
targets file must be named "SAGA_USER_Samples.txt" and must be placed in the same folder together with the sample files.

\cr PLEASE NOTE: The saga package works with raw Agilent data. There is no need for prior data preprocessing.
The saga function will take care of this.
}
