% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_norm.R
\name{saga_norm}
\alias{saga_norm}
\title{Normalize joint SAGA data sets}
\usage{
saga_norm(SAGA_RAW, TEST_RAW, pData.joint, plotnumber = 2)
}
\arguments{
\item{SAGA_RAW}{SAGA raw data matrix}

\item{TEST_RAW}{Test or sample data in SAGA matrix format}

\item{pData.joint}{joint data matrix (internal SAGA data + sample data)}

\item{plotnumber}{default=2 for t-SNE of SAGA model data + sample; alternative = 1 for quantile normalized boxplot with samples}
}
\value{
\code{eset.rma} expression set of normalized array data.
}
\description{
\code{saga_norm} uses both, user data and SAGA data to perform quantile normalization. Multiple probes within arrays are averaged.
}
