% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_predict.R
\name{saga_predict}
\alias{saga_predict}
\title{Prediction of sample arrays with the SAGA model.}
\usage{
saga_predict(samplepath, matrix.train, labels.train, matrix.unknown,
  pData.Test, writeFile = 0, showRoc = 1)
}
\arguments{
\item{samplepath}{path to sample data}

\item{matrix.train}{normalized, probe-averaged and batch-corrected SAGA training data.}

\item{labels.train}{class labels (factors) for SAGA training data. Can either be "transforming" or "nontransforming".}

\item{matrix.unknown}{matrix of sample data with array names as row names and probes as column names.}

\item{pData.Test}{SAGA sample or test data matrix}

\item{writeFile}{default=1 for writing results to the sample folder}

\item{showRoc}{default=1 for showing the ROC curve of the model performance; alternative: 0 for showing naught}
}
\value{
\code{predictions} Data frame with three columns. Column one shows the sample names, the second column shows the decision values of
the svm function and column thress shows the predictions for the query assays either as transforming or nontransforming.
}
\description{
\code{saga_predict} uses a Support Vector Machine (SVM) with a radial kernel to classify user samples either as transforming or nontransforming.
The SVM model is built with the integrated SAGA training data set and a toplist of relevant probes which are able to differentiate assay data
into the mentioned risk states. The SVM is further optimized on the SAGA training data. Use at your own risk!
}
