% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_sampling.R
\name{saga_sampling}
\alias{saga_sampling}
\title{Sample preparation of the joint SAGA data set.}
\usage{
saga_sampling(matrix.SAGA, matrix.test)
}
\arguments{
\item{matrix.SAGA}{Matrix with SAGA model data}

\item{matrix.test}{Matrix with user sample or test data}
}
\value{
\code{matrix.train} normalized, probe-averaged and batch-corrected SAGA training data.

\code{labels.train} class labels (factors) for SAGA training data. Can either be "transforming" or "untransforming".

\code{matrix.unknown} matrix of sample data with array names as row names and probes as column names.
}
\description{
\code{saga_sampling} will prepare training and testing data of the joint SAGA data set. An internal process
is used to filter all data for candidate probes suited for classification of array data into transforming/untransforming.
The list cannot be changed in this version and the package is optimized for this specific gene set. Also, a PCA plot may be
generated to monitor the generalized context of the sample data within the SAGA training set.
}
