% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\docType{data}
\name{sets}
\alias{sets}
\title{sets}
\format{
A 1 x 20 data frame .
}
\usage{
data(sets)
}
\description{
A data frame containing the probe IDs of the top 20 SAGA core features.
}
\examples{
data(sets)
head(sets)
}
\keyword{datasets}
