% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top11.R
\docType{data}
\name{Top11}
\alias{Top11}
\title{Top11}
\format{
A 11 x 2 data frame .
}
\usage{
data(Top11)
}
\description{
Top feature set of the 11 most relevant features.
The data frame contains information on GeneID as well as the corresponding symbol.
}
\examples{
data(Top11)
head(Top11)
}
\keyword{datasets}
