% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top12.R
\docType{data}
\name{Top12}
\alias{Top12}
\title{Top12}
\format{
A 11 x 2 data frame .
}
\usage{
data(Top12)
}
\description{
Top feature set of the 11 most relevant features. Although this file is called Top12, only 11 features are included.
The data frame contains information on GeneID as well as the corresponding symbol.
}
\examples{
data(Top12)
head(Top12)
}
\keyword{datasets}
