% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga_gesea.R
\name{saga_gesea}
\alias{saga_gesea}
\title{Geneset Enrichment Analysis (GESEA) of SAGA data.}
\usage{
saga_gesea(smplpath, saveResults = 0)
}
\arguments{
\item{smplpath}{path to the saga data folder with the user samples.}

\item{saveResults}{can be 1 for yes and 0 for no.}
}
\value{
\code{result} GESEA result - will also be saved into the sample folder.
}
\description{
\code{saga_gesea} performs batch-wise GESEA for SAGA samples.
}
